requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
orjson>=3.9
aiofiles>=23.1.0
google-generativeai>=0.3.2
python-dotenv>=1.0.0
aiosqlite>=0.19.0
pytest-asyncio>=0.23.0
